chore(ast): move statements to seperate file
